Move unsafehacks to separate JAR built separately..
